Corrected the divergence node calculation
The divergence node calculation from earlier doesn't hold because it
assumes that there is only one divergence node. This necessitates a
different (less optimal) method of calculating the first instance of
divergence, which iterates through the shortest path twice (once to get
the set of all nodes, the second time to construct the path to the
divergence node).