Cleaned up outputs some, changed convergence criterion
Expanded the outputs to include the relative gap and a header for the
Beckmann function. The delivery of these results leaves a lot to be
desired and should move away from the list of Doubles that's currently
returned from the optimizer.

Changed the convergence criterion so that it checks for either
completion of maximum number of steps or reaching the desired relative
gap level (if errors in calculating relative gap, fall back to max
iterations).