Abstracted out Link class - new default class is TolledBPRLink
Now that we're beginning to get Links with different travel time
calculation methods we need to abstract out Link so that we can deal
with this effectively. Under the top-level Link is the abstract
TolledLink class, followed by a TolledBPRLink class which includes BPR
method for link travel time calculation.

Next step is to make Link class for COG model