Implement CentroidConnectors, fixed some bugs
CentroidConnectors are now implemented as a subclass of TolledLink (may
move to a TolledCentroidConnector class in the future). Centroid
connectors are checked for validity prior to usage in both the
BushBasedOptimizer.improveBush and Bush.runDijkstras methods so as to
never allow invalid links to be added to the bush. An invalid link is a
centroid connector whose tail node is not the origin.

Added an argument option for specifying nodes which are centroids.

Fixed a bunch of places where BigDecimal division wasn't given a
MathContext to use for rounding repeating decimals. The default context
is now defined as a static value in Optimizer.

Modified Node constructor to define whether a node is a centroid.
Probably should have left an old version of this that didn't need the
second argument, but oh well.

NetworkFactory now handles enhanced networks' centroid connectors (TODO:
provide this for BPR links too).

Added an element comment in Link.alterBushFlow that should hopefully
prevent me (or someone else) in the future from making the stupid
mistake of altering a link's flow directly rather than through the
Bush.changeFlow method (but of course, it's gonna happen again no matter
what).