Converted internal EnhancedLink parameters to Floats
In an effort to cut down on memory overhead, I changed it so all
EnhancedLink attributes are Floats. This should dramatically reduce the
memory overhead for storing the network. This leaves open plenty of
opportunities to improve the computations at a later date by reducing
the BigDecimal arithmetic to Float arithmetic where possible, but for
now the focus is on reducing memory overhead, not computational, so
we're gonna leave it for now.