Removed destDemand field in Bush class to reduce memory overhead
Since the flow on each link from each bush must be stored (or some way
of calculating it must exist), the demand for any bush to any node can
be calculated as the difference in bush flow leaving and entering a
bush. This removes the need to store the demand explicitly, so we can
cut down on memory usage by removing the destDemand field.