consolidated Dijkstra's runs for multiple bushes
Since bushes' initial choice in links is not very consequential, we
allow the origin to run Dijkstra's algorithm to build an initial link
set which will be used initially by all bushes from that origin. Note
that this does NOT guarantee that the links selected by the origin will
be valid for certain bushes, namely when the origin selects links which
do not allow specific vehicle classes.

In this case, on the first iteration of equilibration, these links
should have all flow removed from them. This is accomplished by setting
the toll on links who do not allow the bush's vehicle class to the
maximum float value.