Stable commit for optimizing bush calculations
The previous commits were having issues with optimally determining which
arcs were in bushes, so this has been addressed as follows: Bushes are
now extensions of Graphs (this is a logical extension, since Bushes are
DAGs, which are Graphs). To accomplish this, the Origin of a Bush is no
longer a Node but instead has a Node attribute, and the activeLinks
attribute has been replaced by the attributes contained in Graph.

Additionally, the topoSearch method has been split into two, and Graph
now has the additional methods remove(Link), contains(Link), and
inLinks(Node). The addLink method has been renamed to add(Link)