Re-laid out foundation of main script, expanded trip purposes
Expanded on the data structures and operations of the main script, with
an attempt to make a more understandable data flow path. Lots of things
broken that will need to be rewritten to fit into the new flow, but this
also incorporates all OD matrices into a single script. Will rename
script file on next commit.

Added extra trip purposes to enumeration for future use