A new thread to explore replacing the Enum TripPurpose with a class
This branch will replace the Enums in TripPurpose with instances of the
a TripPurpose class instance (where TripPurpose is an interface). The
goal will be to reduce the spaghetti code by organizing each trip
purpose's operations into a single thread that uses streams more
effectively. Changes in this commit are as follows:

* the old TripPurpose is renamed TripPurposeEnum
* a new interface, TripPurpose, defines the minimum steps provided by a
trip purpose
* a BasicTripPurpose implementation of the TripPurpose interface is
provided that replicates the behavior of the home-based thread and
includes most of the functionality (not including subsegmentation)
* created a basic TripGenerator interface that should be implemented by
all TripGenerators (TODO)
* added a GenerationRate class that will encapsulate both
AreaClass-specific and general trip generation rates (this is very
incomplete)
* TripBalancers should now return a balanced PAMap rather than operating
in place

The following classes are broken and will require significant revision:
* wrapNCTCOG
* ModelInputNCTCOG
* HBThread
* NHBThread
* added an EmptyAggregatePAMatrix class