Changed remaining demand interfaces and classes to use single-precision
This commit gets rid of the lingering instances of double precision in
demand structures and their relevant methods. This method also removes
the AreaSpecificTripGenerator classes and the threads associated with
the former enum-based TripPurpose structure. EnhancedGraphs are also
read in parallel now (improvements still can be made in the concurrency
structure here)