Added feedback capabilities
SkimFactory now can calculate a skim from a provided network and double
function. This depends on FixedSizeNetworkSkim's new dynamic
construction functionality. SkimFactory now also includes the ability to
write a given skim to a given path, but only providing the skim data for
a provided set of zones.

Assigners now provide a skim through the SkimFactory module, provided a
link performance function. As a result, BasicStaticAssigner now keeps
track of its Graph object.

Project now reads in a primary graph, then provides a reference to it
when needed. This is in anticipation of needing to duplicate the network
for feedback loops. Additionally, Assigners are returned as a Map
including their String ids from the properties file. Flow and skim
output after completion is also handled by Project now