Added link-based tolling mechanism for Assigner
This update adds a link-based general tolling strategy to each link in
the network, such that each link adds an additional toll that is defined
by a general network policy in addition to its individual tolls.

The tolling policy is defined in the Assigner, as a lambda function from
a link to its toll. This allows for updating the toll based on any link
characteristic, including (but not limited to) distance,
volume-to-capacity ratio, or endpoints